refactor(merge/split): Rename iob_split2 and iob_merge2. - Deleted old 'iob_split' and 'iob_merge' modules. - Renamed new 'iob_split2' and 'iob_merge2' to 'iob_split' and 'iob_merge', respectively.
